requirements-server.txt
# FastAPI + server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis[hiredis]>=5.0.0

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
google-auth>=2.29.0

# Config
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# AWS (for prod - KMS, Secrets Manager, S3)
boto3>=1.34.0

# Payments
stripe>=8.0.0

# Encryption (Fernet symmetric encryption for vault)
cryptography>=42.0.0

# Utilities
python-ulid>=2.2.0
